Moved the signature generation to GatewayTrait.
